% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypk-package.R
\docType{package}
\name{mypk-package}
\alias{mypk}
\alias{mypk-package}
\title{mypk: refactor codes of a shiny app}
\description{
mypk enable the shiny app to run with simpler code.
}
\author{
\strong{Maintainer}: Junhao Wang \email{jwan0296@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
